Rewrite `stm32f4` setup with svd generated registers
@thalesfragoso this is the patch I mentioned on matrix. Curiously,
@therealprof didn't need it, and you mentioned the current setup has
worked for you in the past as well. Just to clarify, the board I'm
testing is STM32F407G-DISC1 - but I imagine behaviour should be the same
across F407s?

As you mentioned it would need either a critical section or access to
the `&RCC` and `&SYSCFG` passed as arguments to the function directly.

I think I'd opt for the latter approach, and not call this from inside
`Eth::new`, but rather have a separate `configure_rcc` (or something
like this) that the user can call separately before `Eth::new` is
called. If it's a separate function, the user can do

```rust
configure_rcc(&p.RCC, &p.SYSCFG);
```

before they set the RCC clock speeds and call `freeze`. If we really
wanted to make sure this was called prior to `Eth::new`, we could do so
by returning some type from `configure_rcc` that must be passed as an
argument to `Eth::new`.

I'll double check the behaviour of `master` compared to this patch again
tomorrow once I'm back at work on this. I thought I'd open this in the
meantime in case you were curious!